Добавил новые сущности в БД, сделал сервисы, post-запросы на изменение статуса, удаление, добавление блюда.

+++ src/main/java/vsu/netcracker/project/utils/UtilsForAdministrator.java
@@ -1,15 +1,14 @@
 package vsu.netcracker.project.utils;
 
-import javax.persistence.criteria.CriteriaBuilder;
 import java.util.Map;
 
 public class UtilsForAdministrator {
 
-    public static long findKeyWithMaxValueInMap(Map<Long, Integer> map) {
+    public static Integer findKeyWithMaxValueInMap(Map<Integer, Integer> map) {
         int resultVal = 0;
-        long resultKey = 0;
-        for(Map.Entry<Long, Integer> item: map.entrySet()) {
-            if(item.getValue() > resultVal) {
+        Integer resultKey = 0;
+        for (Map.Entry<Integer, Integer> item : map.entrySet()) {
+            if (item.getValue() > resultVal) {
                 resultKey = item.getKey();
                 resultVal = item.getValue();
             }
@@ -17,9 +16,9 @@ public class UtilsForAdministrator {
         return resultKey;
     }
 
-    public static int findMaxValueInMap(Map<Long, Integer> map) {
+    public static int findMaxValueInMap(Map<Integer, Integer> map) {
         int max = 0;
-        for(int val : map.values()) {
+        for (int val : map.values()) {
             if (val > max) max = val;
         }
         return max;
